Add support for individual transforms.
