+++ docs/public/examples/002-draggable-auto-scroll/index.html
@@ -18,6 +18,6 @@
         </svg>
       </div>
     </div>
-    <script type="module" src="index.6ca09920.js"></script>
+    <script type="module" src="index.4d18e56c.js"></script>
   </body>
 </html>
