+++ docs/public/examples/003-draggable-transformed/index.html
@@ -22,6 +22,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="index.11d66023.js"></script>
+    <script type="module" src="index.10b288a7.js"></script>
   </body>
 </html>
