+++ package.json
@@ -39,16 +39,16 @@
     "main": false
   },
   "scripts": {
-    "build": "npm run clean-build && npm run ts-check && npm run build-lib && npm run build-tests && npm run build-examples && npm run format && npm run size",
+    "build": "npm run clean-dist && npm run ts-check && npm run build-lib && npm run build-tests && npm run build-examples && npm run format && npm run size",
+    "build-examples": "node ./scripts/build-examples.mjs",
     "build-lib": "tsup --config ./tsup.config.lib.ts",
     "build-tests": "tsup --config ./tsup.config.tests.ts",
-    "build-examples": "node ./scripts/build-examples.mjs",
-    "clean-build": "rimraf ./dist && rimraf ./tests/dist && rimraf ./docs/public/examples",
+    "clean-dist": "rimraf ./dist && rimraf ./tests/dist && rimraf ./docs/public/examples",
+    "dev": "vitepress dev docs --host",
     "docs-build": "vitepress build docs",
-    "docs-dev": "vitepress dev docs --host",
     "docs-serve": "vitepress serve docs",
-    "lint": "npm run prettier-check && npm run ts-check",
     "format": "npm run prettier-fix",
+    "lint": "npm run prettier-check && npm run ts-check",
     "prettier-check": "prettier --check --no-error-on-unmatched-pattern \"./**/*\"",
     "prettier-fix": "prettier --write --no-error-on-unmatched-pattern \"./**/*\"",
     "size": "for file in ./dist/*.{js,cjs}; do echo $file && gzip-size $file --include-original; done;",
