+++ src/utils/get-world-transform-matrix.ts
@@ -1,3 +1,4 @@
+import { getElementTransformString } from './get-element-transform-string.js';
 import { getStyle } from './get-style.js';
 import { parseTransformOrigin } from './parse-transform-origin.js';
 import { resetMatrix } from './reset-matrix.js';
@@ -14,18 +15,19 @@ export function getWorldTransformMatrix(
   resetMatrix(result);
 
   while (currentElement) {
-    const { transform, transformOrigin } = getStyle(currentElement);
-    if (transform && transform !== 'none') {
-      MATRIX.setMatrixValue(transform);
+    const transformString = getElementTransformString(currentElement);
+    if (transformString) {
+      MATRIX.setMatrixValue(transformString);
       if (!MATRIX.isIdentity) {
+        const { transformOrigin } = getStyle(currentElement);
         const { x, y, z } = parseTransformOrigin(transformOrigin);
         if (z === 0) {
           MATRIX.setMatrixValue(
-            `translate(${x}px, ${y}px) ${MATRIX} translate(${x * -1}px, ${y * -1}px)`,
+            `translate(${x}px,${y}px) ${MATRIX} translate(${x * -1}px,${y * -1}px)`,
           );
         } else {
           MATRIX.setMatrixValue(
-            `translate3d(${x}px, ${y}px, ${z}px) ${MATRIX} translate3d(${x * -1}px, ${y * -1}px, ${z * -1}px)`,
+            `translate3d(${x}px,${y}px,${z}px) ${MATRIX} translate3d(${x * -1}px,${y * -1}px,${z * -1}px)`,
           );
         }
         result.preMultiplySelf(MATRIX);
